+++ RELEASE-NOTES.md
@@ -3,5 +3,6 @@
 * #1 Support for provider methods in modules
 * #2 Support unordered collections (sets)
 * #3 Allow binding of generic types
+* #4 Support for annotations annotated with @Qualifier as injection markers
 * #9 Module superclass with noop implementation of Module.configure(..)
 * #10 Injecting Map with arbitrary key type